% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.isCorrectArray.R
\name{soen6611.isCorrectArray}
\alias{soen6611.isCorrectArray}
\title{Validate an array to contain only numeric.}
\usage{
soen6611.isCorrectArray(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
None
}
\description{
Validate an array to contain only numeric.
}
\examples{
arr <- c(2, 1, 3, 0)
soen6611.isCorrectArray(arr)

soen6611.isCorrectArray(c(2, 1, 3, 0))
}
\keyword{internal}
