% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.max.R
\name{soen6611.max}
\alias{soen6611.max}
\title{Find the largest of the values in the given data set.}
\usage{
soen6611.max(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the largest of the values in the input array
}
\description{
Find the largest of the values in the given data set.
}
\examples{
arr <- <- c(2, 1, 3, 0)
soen6611.max(arr)

soen6611.max(c(2, 1, 3, 0))
}
