% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.mean.R
\name{soen6611.mean}
\alias{soen6611.mean}
\title{Find the mean value in the given data set.}
\usage{
soen6611.mean(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the mean value of the input array
}
\description{
Find the mean value in the given data set.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
soen6611.mean(arr)

soen6611.mean(c(2, 1, 2, 3, 0))

}
