% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.min.R
\name{soen6611.min}
\alias{soen6611.min}
\title{Find the smallest of the values in the given data set.}
\usage{
soen6611.min(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the smallest of the values in the input array
}
\description{
Find the smallest of the values in the given data set.
}
\examples{
arr <- c(2, 1, 3, 0)
soen6611.min(arr)

soen6611.min(c(2, 1, 3, 0))
}
