% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.mode.R
\name{soen6611.mode}
\alias{soen6611.mode}
\title{Find the value that appears most frequently in the given data set.}
\usage{
soen6611.mode(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the value that appears most frequently in the given data set.
        If there is more than one equal mode the smalles will be returned.
}
\description{
Find the value that appears most frequently in the given data set.
}
\examples{
arr <- c(2, 1, 3, 0)
soen6611.mode(arr)

soen6611.mode(c(2, 1, 3, 0))
}
