% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.mode.updated.R
\name{soen6611.mode.updated}
\alias{soen6611.mode.updated}
\title{Find the value that appears most frequently in the given data set.}
\usage{
soen6611.mode.updated(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the value that appears most frequently in the given data set.
        If there is more than one equal mode the smallest will be returned.
}
\description{
Find the value that appears most frequently in the given data set.
}
