% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.quickSort.R
\name{soen6611.quickSort}
\alias{soen6611.quickSort}
\title{Quick sort.}
\usage{
soen6611.quickSort(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
sorted array of numbers
}
\description{
Quick sort.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
sorted_arr <- soen6611.median(arr)

sorted_arr <- soen6611.median(c(2, 1, 2, 3, 0))
}
\keyword{internal}
