% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.sqrt.R
\name{soen6611.sqrt.expcase}
\alias{soen6611.sqrt.expcase}
\title{Function to find the square root of the numbers between 0 and 1}
\usage{
soen6611.sqrt.expcase(number)
}
\arguments{
\item{number}{a number}
}
\value{
square root value
}
\description{
Function to find the square root of the numbers between 0 and 1
}
\examples{
soen6611.sqrt(0.6)
soen6611.sqrt(25)
}
\keyword{internal}
