% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.sqrt.R
\name{soen6611.sqrt}
\alias{soen6611.sqrt}
\title{Find square root of a number}
\usage{
soen6611.sqrt(number)
}
\arguments{
\item{number}{a number}
}
\value{
square root value
}
\description{
Find square root of a number
}
\examples{
soen6611.sqrt(25)
}
