% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.std.R
\name{soen6611.std}
\alias{soen6611.std}
\title{Find the standard deviation value in the given data set.}
\usage{
soen6611.std(input_array)
}
\arguments{
\item{input_array}{input array of numbers}
}
\value{
the standard deviation in the input array
}
\description{
Find the standard deviation value in the given data set.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
soen6611.std(arr)

soen6611.std(c(2, 1, 2, 3, 0))
}
