% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_class.R
\name{Soen6611GroupBClass}
\alias{Soen6611GroupBClass}
\title{Class for Descriptive analitics}
\usage{
Soen6611GroupBClass(input.array = c())
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
Instance of a class for Descriptive analitics
}
\description{
Class for Descriptive analitics
}
\examples{
Soen6611GroupBClass()

arr <- c(2, 1, 3, 0)
Soen6611GroupBClass(arr)

Soen6611GroupBClass(c(2, 1, 3, 0))
}
