% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_max.R
\name{Soen6611Max}
\alias{Soen6611Max}
\title{Find the largest of the values in the given data set.}
\usage{
Soen6611Max(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the largest of the values in the input array
}
\description{
Find the largest of the values in the given data set.
}
\examples{
arr <- c(2, 1, 3, 0)
Soen6611Max(arr)

Soen6611Max(c(2, 1, 3, 0))
}
