% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_mean.R
\name{Soen6611Mean}
\alias{Soen6611Mean}
\title{Find the mean value in the given data set.}
\usage{
Soen6611Mean(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the mean value of the input array
}
\description{
Find the mean value in the given data set.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
Soen6611Mean(arr)

Soen6611Mean(c(2, 1, 2, 3, 0))
}
