% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_median.R
\name{Soen6611Median}
\alias{Soen6611Median}
\title{Find the median value in the given data set.}
\usage{
Soen6611Median(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the median value in the input array
}
\description{
Find the median value in the given data set.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
Soen6611Median(arr)

Soen6611Median(c(2, 1, 2, 3, 0))
}
