% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_min.R
\name{Soen6611Min}
\alias{Soen6611Min}
\title{Find the smallest of the values in the given data set.}
\usage{
Soen6611Min(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the smallest of the values in the input array
}
\description{
Find the smallest of the values in the given data set.
}
\examples{
arr <- c(2, 1, 3, 0)
Soen6611Min(arr)

Soen6611Min(c(2, 1, 3, 0))
}
