% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_mode.R
\name{Soen6611Mode}
\alias{Soen6611Mode}
\title{Find the value that appears most frequently in the given data set.}
\usage{
Soen6611Mode(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the value that appears most frequently in the given data set.
        If there is more than one equal mode the smalles will be returned.
}
\description{
Find the value that appears most frequently in the given data set.
}
\examples{
arr <- c(2, 1, 3, 0)
Soen6611Mode(arr)

Soen6611Mode(c(2, 1, 3, 0))
}
