% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611_std.R
\name{Soen6611Std}
\alias{Soen6611Std}
\title{Find the standard deviation value in the given data set.}
\usage{
Soen6611Std(input.array)
}
\arguments{
\item{input.array}{input array of numbers}
}
\value{
the standard deviation in the input array
}
\description{
Find the standard deviation value in the given data set.
}
\examples{
arr <- c(2, 1, 2, 3, 0)
Soen6611Std(arr)

Soen6611Std(c(2, 1, 2, 3, 0))
}
