% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.quickSort.R
\name{soen6611.merge}
\alias{soen6611.merge}
\title{Merge part of Quick sort.}
\usage{
soen6611.merge(input_array_1, input_array_2)
}
\arguments{
\item{input_array_1}{input array of numbers}

\item{input_array_2}{input array of numbers}
}
\value{
sorted array of numbers
}
\description{
Merge part of Quick sort.
}
\keyword{internal}
