% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soen6611.sqrt.R
\name{soen6611.sqrt}
\alias{soen6611.sqrt}
\title{Function to find square root of numbers}
\usage{
soen6611.sqrt(number)
}
\arguments{
\item{number}{a number}
}
\value{
square root value
}
\description{
Function to find square root of numbers
}
\keyword{internal}
